code refactor

+++ src/handlers/grpc_user.rs
@@ -132,7 +132,7 @@ impl UserService for UserServiceImpl {
                         id: user_id,
                         name: format!("User {}", user_id),
                         email: format!("user{}@example.com", user_id),
-                        age: 30 + (counter % 10) as i32,
+                        age: 30 + (counter % 10),
                     }),
                     update_type: update_type.to_string(),
                     timestamp: timestamp as i64,
